:art: refactors EmeraldButton to reduce number of methods
